backend/requirements.txt
Flask
Flask-Cors
pymongo
dnspython
bcrypt
PyJWT
joblib
requests